ryuan-cli v1.0.2

+++ bin/index.js
@@ -6,22 +6,23 @@ const shell = require('shelljs');
 const log = require('tracer').colorConsole()
 
 program
-    .version('1.0.0')
-    .description('ryuan前端模板工程的cli')
+    .version('1.0.2','-v, --version')
+    .description('This is a front - end template that can quickly generate front-end projects.')
+
 program
-    .command('* <project>')
-    .action(function(project) {
-        // log.info('目前ryuan-cli支持以下模板：')
-        log.info('使用例子：ryuan-cli <project>')
-        if (project) {
+    .command('ryuan-cli <tpl> <project>','tpl:gulp or wevpack')
+    .action(function(tpl , project) {
+        log.info('example：ryuan-cli <tpl> <project-name>')
+        if (tpl && project) {
+            // console.log(tpl);
             let pwd = shell.pwd()
-            log.info(`正在拉取模板代码，下载位置：${pwd}/${project}/ ...`)
-            clone(`https://github.com/Ry-yuan/activity-project.git`, pwd + `/${project}`, null, function() {
+            log.info(`The template code is being pulled.Download location:${pwd}/${project}/ ...`)
+            clone(`https://github.com/Ry-yuan/${tpl}-workflow`, pwd + `/${project}`, null, function() {
                 shell.rm('-rf', pwd + `/${project}/.git`)
-                log.info('模板工程项目建立完成')
+                log.info('The formwork project is completed.Start your work!')
             })
         } else {
-            log.error('正确命令例子：ryuan-cli myproject')
+            log.info('The correct command example：ryuan-cli gulp projectName')
         }
     })
 program.parse(process.argv)
